enshrouded/requirements.txt
apprise
pyyaml
watchdog
subprocess
requests
